moved prefetch sw components under buffering
