+++ packages/qwik/src/core/buffering/components/prefetch.ts
@@ -1,9 +1,9 @@
 // keep this import from qwik/build so the cjs build works
 import { isDev } from '@builder.io/qwik/build';
 import type { JSXNode } from '@builder.io/qwik/jsx-runtime';
-import { _jsxC } from '../internal';
-import type { JSXOutput } from '../render/jsx/types/jsx-node';
-import { useServerData } from '../use/use-env-data';
+import { _jsxC } from '../../internal';
+import type { JSXOutput } from '../../render/jsx/types/jsx-node';
+import { useServerData } from '../../use/use-env-data';
 
 /**
  * Install a service worker which will prefetch the bundles.
